device agnostic refactoring to V1 model runner
